+++ src/auth/LoginButton.vue
@@ -1,16 +1,17 @@
 <template>
-  <a :href="getLoginUrl(route.query.destination || '/corpus')">
-    <ActionButton variant="primary">
-      <icon :icon="['fas', 'right-to-bracket']" />
-      {{ $t("login") }}
-    </ActionButton>
-  </a>
+  <UrlButton
+    class="button-primary"
+    :href="getLoginUrl(route.query.destination || '/corpus')"
+  >
+    <icon :icon="['fas', 'right-to-bracket']" />
+    {{ $t("login") }}
+  </UrlButton>
 </template>
 
 <script setup>
 import { useRoute } from "vue-router";
 import { getLoginUrl } from "./auth";
-import ActionButton from "@/components/ActionButton.vue";
+import UrlButton from "@/components/UrlButton.vue";
 
 const route = useRoute();
 </script>