+++ src/corpus/corpus.composable.js
@@ -4,7 +4,6 @@ import { useCorpusStore } from "@/store/corpus.store";
 import useMessenger from "@/message/messenger.composable";
 import useCorpora from "@/corpora/corpora.composable";
 import useConfig from "./config/config.composable";
-import useExports from "./exports/exports.composable";
 
 /** Let data be refreshed initially, but skip subsequent load calls. */
 const isCorpusFresh = {};
@@ -13,10 +12,9 @@ export default function useCorpus(corpusId) {
   const corpusStore = useCorpusStore();
   const { refreshJwt } = useAuth();
   const mink = useMinkBackend();
-  const { loadCorpora } = useCorpora();
+  const { loadCorpora, refreshCorpora } = useCorpora();
   const { alertError } = useMessenger();
   const { loadConfig } = useConfig(corpusId);
-  const { loadExports } = useExports(corpusId);
 
   async function loadCorpus(force = false) {
     // Make sure the corpus has an entry in the store.
@@ -25,23 +23,14 @@ export default function useCorpus(corpusId) {
       return;
     }
 
-    // Load all essential info about the corpus.
-    await Promise.all([
-      loadConfig(), //
-      loadExports(),
-      loadResourceInfo(),
-    ]);
+    // Load remaining essential info about the corpus.
+    // Skip if removed.
+    if (corpusId in corpusStore.corpora) {
+      await loadConfig();
+    }
 
     // Remember to skip loading next time.
     isCorpusFresh[corpusId] = true;
-  }
-
-  /** Load job status and source files in the same request. */
-  async function loadResourceInfo() {
-    const info = await mink.resourceInfo(corpusId).catch(alertError);
-    corpusStore.corpora[corpusId].name = info.resource.name;
-    corpusStore.corpora[corpusId].sources = info.resource.source_files;
-    corpusStore.corpora[corpusId].status = info.job;
   }
 
   async function deleteCorpus(corpusId_ = corpusId) {
@@ -50,7 +39,7 @@ export default function useCorpus(corpusId) {
     // The backend will have updated the remote JWT, so refresh our copy.
     // The backend uses the corpus list within it when listing available corpora.
     await refreshJwt();
-    corpusStore.removeCorpus(corpusId_);
+    await refreshCorpora();
   }
 
   return { loadCorpus, deleteCorpus };