+++ src/home/Home.vue
@@ -4,8 +4,10 @@ import Section from "@/components/Section.vue";
 import { getLogoutUrl } from "@/auth/auth";
 import LoginButton from "@/auth/LoginButton.vue";
 import ActionButton from "@/components/ActionButton.vue";
+import UrlButton from "@/components/UrlButton.vue";
 import HomeIllustration from "./HomeIllustration.vue";
 import HomeNews from "./HomeNews.vue";
+import RouteButton from "@/components/RouteButton.vue";
 
 const { isAuthenticated, canUserWrite, payload } = useAuth();
 const logoutUrl = getLogoutUrl();
@@ -40,12 +42,10 @@ const logoutUrl = getLogoutUrl();
             </div>
 
             <div v-if="!isAuthenticated">
-              <router-link to="/signup">
-                <ActionButton variant="success">
-                  <icon :icon="['fas', 'user-plus']" />
-                  {{ $t("signup") }}
-                </ActionButton>
-              </router-link>
+              <RouteButton to="/signup" class="button-success">
+                <icon :icon="['fas', 'user-plus']" />
+                {{ $t("signup") }}
+              </RouteButton>
               <div class="my-1 text-sm opacity-70">
                 {{ $t("signup.help") }}
               </div>
@@ -57,18 +57,14 @@ const logoutUrl = getLogoutUrl();
             >
               <div>{{ $t("welcome", { name: payload.name }) }}</div>
 
-              <router-link to="/corpus">
-                <ActionButton variant="primary">
-                  {{ $t("mydata") }}
-                </ActionButton>
-              </router-link>
+              <RouteButton to="/corpus" class="button-primary">
+                {{ $t("mydata") }}
+              </RouteButton>
 
-              <a :href="logoutUrl">
-                <ActionButton>
-                  <icon :icon="['fas', 'person-running']" />
-                  {{ $t("logout") }}
-                </ActionButton>
-              </a>
+              <UrlButton :href="logoutUrl">
+                <icon :icon="['fas', 'person-running']" />
+                {{ $t("logout") }}
+              </UrlButton>
             </div>
           </div>
         </div>