+++ src/user/admin.composable.js
@@ -8,23 +8,28 @@ const adminModeRef = ref(false);
 
 export default function useAdmin() {
   const { canUserAdmin } = useAuth();
-  const { loadCorpora } = useCorpora();
+  const { refreshCorpora } = useCorpora();
   const mink = useMinkBackend();
   const { alertError } = useMessenger();
+
+  async function checkAdminMode() {
+    adminModeRef.value = await mink.checkAdminMode().catch(alertError);
+  }
 
   async function enableAdminMode() {
     await mink.enableAdminMode().catch(alertError);
     adminModeRef.value = true;
-    await loadCorpora(true);
+    await refreshCorpora();
   }
 
   async function disableAdminMode() {
     await mink.disableAdminMode().catch(alertError);
     adminModeRef.value = false;
-    await loadCorpora(true);
+    await refreshCorpora();
   }
 
   return {
+    checkAdminMode,
     enableAdminMode,
     disableAdminMode,
     adminMode: readonly(adminModeRef),