+++ CHANGELOG.md
@@ -13,6 +13,30 @@ As this project is a user-facing application, the places in the semantic version
 
 ## [Unreleased]
 
+## [1.2.0] (2024-01-17)
+
+### Added
+
+- Option to enable named entity recognition
+- Check if admin mode is enabled on load
+
+### Changed
+
+- Dependencies upgraded
+  - Node 18 is now required, due to @vitejs/plugin-vue@5
+- Request `resource-info` once for all resources
+- `UrlButton` component replaces `a > ActionButton`
+- `RouteButton` component replaces `router-link > ActionButton`
+- `ActionButton` uses `<button>` and is tabbable
+
+### Fixed
+
+- Error handling when deleting corpus
+
+### Removed
+
+- `useVariant`
+
 ## [1.1.0] (2024-01-02)
 
 ### Added
@@ -99,7 +123,8 @@ The frontend is now open to the general public! This version allows users to:
 
 Code changes up until this point are not documented other than in the git commit log.
 
-[unreleased]: https://github.com/spraakbanken/mink-frontend/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/spraakbanken/mink-frontend/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/spraakbanken/mink-frontend/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/spraakbanken/mink-frontend/compare/v1.0.5...v1.1.0
 [1.0.5]: https://github.com/spraakbanken/mink-frontend/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/spraakbanken/mink-frontend/compare/v1.0.3...v1.0.4